templating domain flow for loading image feed
`LoadFeedInteractor` is the connection between the `back end` and the presentation layer, it contains logic about where to get the data from and how it is retrieved and processed.
`LoadFeedOutput` is implemented by the caller and the output of the interactor is returned through  the implemented methods on the output interface.
`FeedRemoteGateway` is a remote gateway interface defining the contract for communication between the domain and the remote modules. Implementation is hidden behind the method definitions available in the interface.