simple request class for public feeds
requests to remote connections may consist of a path, query params, body, headers, etc. This is the start of an object that removes knowledge of the content of the request into a single place. Can be extended to dynamically add query parameters through a builder instance, e.g. searching by tags.